Separate import and Maxfilter steps
Skip import if we don't rename events, add or remove
channels, or wish to find flats or bads via Maxwell-filtering.
